+++ CHANGES.md
@@ -2,6 +2,15 @@
 
 ---
 
+# v0.2.0
+
+### New
+
+- Install the Strimzi Kafka operator;
+- Provision an ephemeral kafka broker in Kubernetes.
+
+---
+
 # v0.1.0
 
 ### Maintenance